refactored code to add other oauth providers
